refactor: rename 'NonORM' to 'NoStruct'
