+++ src/DataHandler/DataHandlerManagerInterface.php
@@ -2,11 +2,11 @@
 
 namespace CraftyDigit\Puff\DataHandler;
 
-use CraftyDigit\Puff\DataHandler\NonORM\NonORMEntityManagerInterface;
+use CraftyDigit\Puff\DataHandler\NoStruct\NoStructEntityManagerInterface;
 use CraftyDigit\Puff\Enums\DataHandler;
 use Doctrine\ORM\EntityManagerInterface;
 
 interface DataHandlerManagerInterface
 {
-    public function getEntityManager(DataHandler $dataHandler): NonORMEntityManagerInterface|EntityManagerInterface;
+    public function getEntityManager(DataHandler $dataHandler): NoStructEntityManagerInterface|EntityManagerInterface;
 }
