+++ src/DataHandler/NoStruct/NoStructEntityManagerInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace CraftyDigit\Puff\DataHandler\NoStruct;
+
+interface NoStructEntityManagerInterface
+{
+    public function getRepository(string $dataSourceName): NoStructRepositoryInterface;
+}
