--- src/DataHandler/NonORM/AbstractNonORMRepository.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace CraftyDigit\Puff\DataHandler\NonORM;
-
-abstract class AbstractNonORMRepository implements NonORMRepositoryInterface
-{
-    public function __construct(protected readonly string $dataSourceName)
-    {}
-}
