--- src/DataHandler/NonORM/NonORMEntityManagerInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace CraftyDigit\Puff\DataHandler\NonORM;
-
-interface NonORMEntityManagerInterface
-{
-    public function getRepository(string $dataSourceName): NonORMRepositoryInterface;
-}
